Make identities reentrant to UID change
This change makes it so that if an Identity references a user with
the wrong UID or a non-existent user, that identity is deleted
during provisioning. This UID mismatch is caused by the deletion and
possible recreation of users.  Thus this makes it so that the "new"
user is unaffected by the "old" user's identity.  The new user
simply goes through the provisioning just as if no identity existed.

Signed-off-by: Monis Khan <mkhan@redhat.com>