Correctly handle newlines in SerialFileGenerator
This change makes it so that SerialFileGenerator correctly reads and
writes serial files that end in a newline.  This allows it to
interoperate with other tools that may interact with the serial file
such as openssl.  Tests were added to assert that the incrementing
logic works and that the serial file always ends with a newline.

Bug 1512825

Signed-off-by: Monis Khan <mkhan@redhat.com>