Allow authorization of SCC use via RBAC
This change adds the minimal logic required to support the following
SAR check as a means to allow access to a specific SCC in a specific
namespace:

Verb:      use
Namespace: <namespace>
Name:      <scc name>
APIGroup:  security.openshift.io
Resource:  securitycontextconstraints

This SAR check mimics the upstream PSP check but with the SCC group
and resource.  This has the added benefit of allowing namespace
scoped SCC access.

It does not update any bootstrap data to use this method.  It does
not update the router command's SCC check to use this method.  Thus
nothing currently relies on this method as a means of SCC access.

Signed-off-by: Monis Khan <mkhan@redhat.com>