Allow empty domain in AcquireUserCredentials
This change allows the use of the empty string as the domain
parameter to AcquireUserCredentials.  Based on my testing with a
Windows 2016 server, domain can be left empty as long as the
username is qualified in either user@foo or user@foo.domain.tld
format (user and foo\user does not work).

Signed-off-by: Monis Khan <mkhan@redhat.com>