Update GitLab IDP to use OIDC instead of OAuth2
This change updates the GitLab IDP integration to use OpenID Connect
instead of OAuth2.  This has the benefit of using a standard that
was meant for authentication, which allows us to remove the GitLab
specific code from our integration.  We can simply reuse the OIDC
code with a GitLab specific configuration.  From the perspective of
the end user, nothing changes.  The configuration of the IDP remains
exactly the same and the identity/user objects are unchanged.

The only difference is that a very recent version of GitLab is
required: version 11.1.0 or later.

A positive side effect from this change is that it allows us to use
the openid scope instead of the api scope.  This means that
OpenShift only gets limited read access to perform authentication
instead of full read/write access to the user's GitLab account.  The
user may have to reauthenticate OpenShift's OAuth client on first
login.

In the future this change will make it easy to pull group membership
information from GitLab (the tokens already include a groups claim,
we simply need to add code in OpenShift that uses it).

Trello: https://trello.com/c/DXntmEOV

Signed-off-by: Monis Khan <mkhan@redhat.com>