Remove misleading CreationTimestamp from tokens
The OSIN storage code gives the false appearance of having control
over the creation timestamp of API objects (in this case
OAuthAuthorizeToken and OAuthAccessToken).  While this timestamp
will be close to the value stored in OSIN data structures, it will
not be the same since the tokens are created afterwards.

Signed-off-by: Monis Khan <mkhan@redhat.com>