Refactor OAuth session logic
The change significantly refactors the OAuth session code.  The
important highlights of this change are:

1. Remove duplicate Mux interface logic.
2. Remove session wrapping logic.  This was used to clean up global
state but is no longer required (see below).
3. Change session cookie logic to never error on cookie decoding.
Such errors are usually caused by poor key rotation and there is no
reason to require the user to restart their browser or otherwise
clear the session cookie to log in again.
4. Simplify the session interfaces to significantly improve the
readability of the session authenticator.
5. Have the session authenticator store the expiration information
as an int64 in the cookie.  This allows us to skip the conversion
from a string back into an integer.  The old string field is
preserved to allow for mixed master scenarios.
6. Change the session cookie logic to not maintain global state per
request.  Consequently, no clean up of global state is required.
This prevents a host of difficult to debug memory leaks [1][2][3].

[1] https://github.com/gorilla/sessions/issues/122
[2] https://github.com/gorilla/sessions/pull/110
[3] https://github.com/gobuffalo/buffalo/pull/619

This change also removes the OAuth server's direct dependency on
github.com/gorilla/securecookie and github.com/gorilla/context.

Signed-off-by: Monis Khan <mkhan@redhat.com>