Include all groups in ~ user
This change updates the virtual ~ user to include all groups on the
request (irregardless of if the user is virtual or not).  This makes
it so that the API now mirrors exactly what the server sees for
authentication purposes.

Similarly, the UID from the request is shown in favor of the UID on
any persisted user API object (since that is what authentication
will use to make decisions).

There currently is still no way for the user to see what extra
parameters were applied to the request.

The primary reason for this change is to allow for better
introspection with remote token authentication and with groups from
identity providers that are not reflected in OpenShift groups.  It
also allows unprivileged users to discover their full group
membership even when that membership is solely based on OpenShift
groups (since normal users cannot see the groups API objects).

Signed-off-by: Monis Khan <mkhan@redhat.com>