Consolidate user conversion logic
This change removes the duplication of logic in the user conversion
code and tightens the code to avoid interface{} where possible.  It
also enforces that a token must always have a UserName and UserUID
(which is always the case since it is enforced via validation).

To make things simpler, the UserConversion interface was removed and
the logic was turned into an implementation detail of our custom
osin.Storage.  The OAuthStorage integration test was significantly
updated to accommodate this change and to make it easier for future
readers to determine what the test actually asserts and how it is
configured.

Signed-off-by: Monis Khan <mkhan@redhat.com>